added hp bar

+++ Assets/Project/Scripts/Gameplay/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageable
 {
-    void ApplyDamage(float damage);
+    void ApplyDamage(int damage);
 }
